fix: betterer problems, use seperate tsconfig for build process and use strict mode for IDE
