+++ .betterer.ts
@@ -20,8 +20,5 @@
 import {typescript} from '@betterer/typescript';
 
 export default {
-  'stricter compilation': () =>
-    typescript('./tsconfig.json', {
-      strict: true,
-    }).include('./src/**/*.{ts,tsx}'),
+  'stricter compilation': () => typescript('./tsconfig.json').include('./src/**/*.{ts,tsx}'),
 };
