+++ .eslintrc.json
@@ -21,7 +21,7 @@
       "files": ["*.ts", "*.tsx"],
       "parser": "@typescript-eslint/parser",
       "parserOptions": {
-        "project": ["./tsconfig.json", "./server/tsconfig.json"]
+        "project": ["./tsconfig.build.json", "./server/tsconfig.json"]
       },
       "rules": {
         "@typescript-eslint/no-floating-promises": "off",
