+++ server/package.json
@@ -37,7 +37,7 @@
     "typescript": "4.7.4"
   },
   "scripts": {
-    "build": "yarn clean && tsc -P tsconfig.json && yarn copy-assets && yarn create-commit-sha",
+    "build": "yarn clean && tsc -P tsconfig.build.json && yarn copy-assets && yarn create-commit-sha",
     "copy-assets": "node ./bin/copy_server_assets.js",
     "create-commit-sha": "node ./bin/create_commit_sha_file.js",
     "clean": "rimraf ./dist/",
