+++ src/script/main/globals.ts
@@ -1,6 +1,6 @@
 /*
  * Wire
- * Copyright (C) 2019 Wire Swiss GmbH
+ * Copyright (C) 2022 Wire Swiss GmbH
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -18,129 +18,7 @@
  */
 
 import {amplify} from 'amplify';
-import jQuery from 'jquery';
 import ko from 'knockout';
-import type {t} from 'Util/LocalizerUtil';
-
-import type {AssetService} from '../assets/AssetService';
-import type {NotificationService} from '../event/NotificationService';
-
-import '../Config';
-
-import 'Components/CopyToClipboard';
-import 'Components/UserSearchableList';
-import 'Components/toggle/BaseToggle';
-import 'Components/icons';
-import 'Components/Image';
-import 'Components/toggle/InfoToggle';
-import 'Components/calling/FullscreenVideoCall';
-import 'Components/avatar/GroupAvatar';
-import 'Components/MessagesList';
-import 'Components/TitleBar';
-import 'Components/Modals/GroupCreation/GroupCreationModal';
-import '../page/AppLock';
-import 'Components/modal';
-import 'Components/panel/ServiceDetails';
-import 'Components/panel/UserDetails';
-import 'Components/ServiceList';
-import 'Components/SearchInput';
-import 'Components/calling/CallingOverlayContainer';
-
-import 'Util/LocalizerUtil';
-
-import '../localization/Localizer';
-import '../view_model/bindings/CommonBindings';
-import '../view_model/bindings/ConversationListBindings';
-import '../view_model/bindings/MessageListBindings';
-import '../view_model/bindings/VideoCallingBindings';
-
-import '../view_model/MainViewModel';
-import '../view_model/PanelViewModel';
-
-declare global {
-  interface Window {
-    $: typeof jQuery;
-    amplify: amplify.Static;
-    jQuery: typeof jQuery;
-    ko: typeof ko;
-    t: typeof t;
-    wire: {
-      app: {
-        service: {
-          asset: AssetService;
-          notification: NotificationService;
-        };
-      };
-      env: {
-        ANALYTICS_API_KEY: string;
-        APP_BASE: string;
-        APP_NAME: string;
-        BACKEND_NAME: string;
-        BACKEND_REST: string;
-        BACKEND_WS: string;
-        BRAND_NAME: string;
-        CHROME_ORIGIN_TRIAL_TOKEN: string;
-        COUNTLY_API_KEY: string;
-        ENVIRONMENT: string;
-        FEATURE: {
-          ALLOWED_FILE_UPLOAD_EXTENSIONS: string[];
-          APPLOCK_SCHEDULED_TIMEOUT: number;
-          CHECK_CONSENT: boolean;
-          CONFERENCE_AUTO_MUTE: boolean;
-          DEFAULT_LOGIN_TEMPORARY_CLIENT: boolean;
-          ENABLE_ACCOUNT_REGISTRATION: boolean;
-          ENABLE_ACCOUNT_REGISTRATION_ACCEPT_TERMS_AND_PRIVACY_POLICY: boolean;
-          ENABLE_DEBUG: boolean;
-          ENABLE_DOMAIN_DISCOVERY: boolean;
-          ENABLE_ENFORCE_DESKTOP_APPLICATION_ONLY: boolean;
-          ENABLE_EXTRA_CLIENT_ENTROPY: boolean;
-          ENABLE_MEDIA_EMBEDS: boolean;
-          ENABLE_MLS: boolean;
-          ENABLE_PHONE_LOGIN: boolean;
-          ENABLE_SSO: boolean;
-          ENFORCE_CONSTANT_BITRATE: boolean;
-          PERSIST_TEMPORARY_CLIENTS: boolean;
-          SHOW_LOADING_INFORMATION: boolean;
-        };
-        MAX_GROUP_PARTICIPANTS: number;
-        MAX_VIDEO_PARTICIPANTS: number;
-        NEW_PASSWORD_MINIMUM_LENGTH: number;
-        URL: {
-          ACCOUNT_BASE: string;
-          MOBILE_BASE: string;
-          PRICING: string;
-          PRIVACY_POLICY: string;
-          SUPPORT: {
-            BUG_REPORT: string;
-            CALLING: string;
-            CAMERA_ACCESS_DENIED: string;
-            CONTACT: string;
-            DEVICE_ACCESS_DENIED: string;
-            DEVICE_NOT_FOUND: string;
-            EMAIL_EXISTS: string;
-            HISTORY: string;
-            INDEX: string;
-            LEGAL_HOLD_BLOCK: string;
-            MICROPHONE_ACCESS_DENIED: string;
-            PRIVACY_VERIFY_FINGERPRINT: string;
-            SCREEN_ACCESS_DENIED: string;
-          };
-          TEAMS_BASE: string;
-          TEAMS_BILLING: string;
-          TEAMS_CREATE: string;
-          TERMS_OF_USE_PERSONAL: string;
-          TERMS_OF_USE_TEAMS: string;
-          WEBSITE_BASE: string;
-          WHATS_NEW: string;
-        };
-        VERSION: string;
-        WEBSITE_LABEL: string;
-      };
-    };
-    wSSOCapable: boolean;
-    z: any;
-  }
-}
 
 window.amplify = amplify;
 // we need to publish jQuery on the window so that knockout can use it
